Moved map output to a temp directory.
